+++ Source/Modules/xISHInstall/xISHInstall.psm1
@@ -1,0 +1,5 @@
+#Requires –RunAsAdministrator
+
+. $PSScriptRoot\Install-ISHDeployment.ps1
+. $PSScriptRoot\New-ISHDeploymentInputParameters.ps1
+. $PSScriptRoot\Uninstall-ISHDeployment.ps1